replace recursive vector with global array, it was unnecessary!
I am not sure why I thought the overwriting was a problem.
It does not run in parallell. When forward returns it always goes just one
step backwards, and then when forwarding again it only overwrites
indices that are no longer needed.

I have reverted to a global array which means that memory is a non-issue.

I now have from Katniss: Time Limit Exceeded (2/2)
This is the furthest I've gotten so far!
So this means that the algorithm itself must be more effective.